researching what objects in the k8s API we'll actually want to watch like a hawk.
It's certainly less than all of them -- events, etc, of course are not to be declaratively created.  But it's most of them.

There seems to be a fairly serious issue at the heart of the k8s designs: it's not *entirely* clear-cut which objects are roots of ownership trees.  There seems to be work towards this being described, but it's... still lacking clear roots and ways to opt-in to ownership models clearly, and thus overall the model has seriously problematic gaps.

Signed-off-by: Eric Myhre <hash@exultant.us>