hello-world controller loop.
This is based largely on the demonstration content from https://github.com/kubernetes/community/blob/master/contributors/devel/controllers.md (which is also Apache licensed, natch).

I applied a jsonpatch library used elsewhere in the k8s internals to try to print out terse status diffs out of interest in what we'd get.

I'm still a little tepid on the library levels used here.  I want to avoid leaning on concrete type definitions too much because that's basically an opportunity to lose sight of data if the cluster has anything other than the exact same version of structures as we do... so there's some work to do in getting *back* to basics here.

Signed-off-by: Eric Myhre <hash@exultant.us>