Some commentary on basic desires for a dashboard.
The kubectl command suffers from Conway's law poignantly:
all of its operations mirror the internal abstractions
of kubernetes quite clearly, and in doing so, it almost
completely misses the mark on the needs of the users.

Concretely: if you understand the kubernetes object model,
it's perfectly understandable why `kubectl get pods`
eschews much info about statuses and when they changed.
And yet it's completely bonkers to a user; we *want* a
join with the events table to know when statuses occured
pretty much 100% of the time we're looking at pods.

Signed-off-by: Eric Myhre <hash@exultant.us>