Replace that entire cacher monstrosity with plain watch.
I have... no idea... why that cacher thing is 1/80th as complicated as it is.

As described in the previous commit, though, we're still going to try to
go *even simpler* than this: it is actively preferable *not* to map things
into the type system, both because it may lose fields if we don't have
a perfect version match, and because we'd quite simply end up using *more
reflection* in order to iterate things again, and that's *going backwards*.

Unfortunately, the k8s API client libraries are *not* making that easy on us.
More research todo.

Signed-off-by: Eric Myhre <hash@exultant.us>