Print func for UnstructuredList type, just to get myself traversing it.
The actual objects are, as usual, about one layer further indirection
away from you than makes any sense.

(I'm particularly triggered by the `EachListItem` method.  This isn't
javascript, and we don't have generics either, so this is just about
the *worst* possible thing you could do to yourself in Golang!  It takes
a closure parameter of course... which is `func(runtime.Object) error`,
which says approximately *nothing* useful and almost certainly leaves
you in a casting hell.  Why you would do this over simply *iterating
the dang list* is completely unfathomable to me (doubly so in the dark
twisty corridors of k8s world where codegen is thrown about as freely
as rum on a pirate ship, and there's no possible argument for *not*
simply *generating* for loops if you really got flustery about it).
And then when you get the `unstructured.Unstructured`, literally the
only thing you would ever do with this is deference the `.Object` field,
which is `map[string]interface{}`... so, really: ALL of this could
have been `[]map[string]interface{}` and we'd be just as well off.
Whee!)

Signed-off-by: Eric Myhre <hash@exultant.us>