Discovery client package is what we want.
So... nice!  This is simply just what I wanted, for once.

It *is* kind of slow, though.  This API is just a comedy of REST overhead.

Just getting the groups list: real:0.25 user:0.04 sys:0.00, for example.

Getting the actual itemized resources: real:1.41 user:0.05 sys:0.01.

Long story short: yeah, every time you feel like there's almost a second
and a half of totally unreasonable network wait at the beginning of your
`kubectl` commands?  *You're not imagining it.*  It's real.  And no,
the reason for it is *not* sensible.

Signed-off-by: Eric Myhre <hash@exultant.us>