I was wrong.  The discovery package is only very remotely a comic impression of what we want.
See the massive documentation comment on the 'saneDiscovery' method.
tl;dr, the structuring of the data we receive from the discovery package
is a parody of anything you're actually likely to want to work with.

But, hey, we've finally got something.  I rigged it to our list call
for deployments, and it does the thing; hardcoded version string gone.

Comments about multiple groups here are prompted because I noticed
that Deployment is *actually doing that*, which scares the pants off
of me.  I thought it was canonically considered under the "extensions"
group; turns out, it's *also* listed under the "apps" group.  Wow!

Signed-off-by: Eric Myhre <hash@exultant.us>