Make method to hide most of the redundant futz needed to set up making a single call to a resource.
Almost all of this is redundant massaging of config objects.
All of the possible error paths are just checks from many different layers on the validity
of said config.

The happy news is the structure we get out of this does actually support doing e.g. both
"Get" and "List" and "Watch" and so on, which is roughly what we want.

Note though that it's no coincidence that this diff is only replacing the deployments list demo!
This approach still doesn't work yet to replace the namespace list.  Why?  Stay tuned,
for our next commit in K8s API Comedy Theater 3000!

Signed-off-by: Eric Myhre <hash@exultant.us>