Slight simplification to avoid making another APIResource object.
Not that it matters:

Try to comment out that `&schema.GroupVersion` allocation and string reshuffle.
It looks like the most redundant thing in the world, right?  Yeah, well.
It doesn't work without that.  Because the `Resource` method, in its infinite
wisdom, actually strips out and *ignores* every part of the `APIResource` object...
except the Name.

So, is *everything about* that method signature a joke?  Yes.  Yes it is.

Signed-off-by: Eric Myhre <hash@exultant.us>