Support dynamic listing for 'legacy' 'core' (yes, BOTH of those words describe the same thing; 'legacy' here meaning keep it forever but bless its weirdness) resources.
"/api" and "/apis".  Why not?  This *definitely* won't generate any
cognitive load for API users, or be likely to pile up with other cognitive
load generators into a giant heaping logjams of confusing details that causes
developers to waste hours when they're trying to pursue the most basic
proof-of-concept uses of the API...

... oh wait we have an experience report of EXACTLY THAT in commit
8e91b526d539cc6f4431e90b50319329c34eae6a

It's almost like complexity is this thing you can measure, and
it's almost like complexity has predictable impacts on productivity...

Signed-off-by: Eric Myhre <hash@exultant.us>