Finally: bring our simplified client in for Watch.
This works as you would expect, and at long last, we can watch arbitrary
resources and get simple traversable maps and lists back out.

Wow.

It only took... a month... to get the couple of basic "$thing, $namespace: watch"
calls that I wanted.

We're importing fewer packages than ever, as of this commit.
The trouble was entirely in figuring out *which handful* of packages
are the *right* ones for actually doing something simple.

Signed-off-by: Eric Myhre <hash@exultant.us>