do some shorter printing of changes.
I'm poking around with the idea that maybe it would be useful to have a
k8s-tail-all tool.  It's unclear how to make this *useful* though -- it's
entirely capable of producing vastly more torrential amounts of information
than can possibly be used.  But, hey, maybe piping that to a grep will
be what you want sometimes?

It looks like some watch calls go futher back in history than others, which
is interesting.  e.g. the namespace watch seems to effectively list all of
them, whereas the events watch does not appear to be so backlogging.

Resource versions are not global, I also notice.  So this is going to make
it tricky (read: impossible?) to do consistent read snapshots across
categories of resources.  Hrrmph.

Signed-off-by: Eric Myhre <hash@exultant.us>