Filter files by extension.
This immediately fixes some of our issues in prod (if in ways which are the
worst connotations of "practical"), but also not all.

Have I ever mentioned I really don't think git-filter configs are a good idea?
They mean you never actually know if that path contains data you can *use or not*.

Signed-off-by: Eric Myhre <hash@exultant.us>