More printouts; powered by templates.
The theory I'm running with here for the moment is that we should be able to
put these "Perspective" objects in a json config file... and understand
arbitrary objects without recompiling rekkoner: across k8s versions, freely
reckoning with third-party extension objects, and so on.

And thus to do that... all those fields need to be serializable.

And templates are strings!

So far so good.

Looking at what happens in practice though, I think we might actually want
to pay plenty of attention to the filenames that were the sources of what
we loaded.  *Those* are actually often the best names of all for communicating
what part of the *end user's* ontologies we're flailing around.

Signed-off-by: Eric Myhre <hash@exultant.us>