holy smokes aliases are awesome.
I am going to be *so* very pleased to be free of needing to import what seems
to be 99 almost randomly named k8s package trees for basic operations.

(Seriously.  My IDE can't even guess the right ones half the time because
there's just too many and their names are too ambiguous and overlappy.  This
handful of aliases is going to save so, so much headache.)

Signed-off-by: Eric Myhre <hash@exultant.us>