Quick printoff of objects in cluster.
Not diffing/joining versus the intents yet, but simply from the fact it's
using the same set of Kinds, we get to start observing some interesting things.

(For starters: is the number of objects in the prod I'm examining right now
even within 2x of the number of config files?  Nope! :D  Not even 3x.  More!
*This is **why** we need this tool.*)

Signed-off-by: Eric Myhre <hash@exultant.us>