Change to naming strategy.
Pretty sure the namespace::kind::name tuple is correct 99% of the time.
The generate-name might come into play somewhere, but iterating on that
in the future as I figure out how.

I really would like to declare that using the namespace as an *argument* at
'kubectl apply' time is such an antipattern I'm going to engineer as if it
isn't allowed.  Doing so really screws with our idea of serialized intent.
(Maybe we can bake this back in via more flags later?  But it's really
a nasty break in ergonomics if you have to turn around and run rekkoner
repeatedly for multiple namespaces, since remember, some things *aren't
namespaced* in the first place.)

Signed-off-by: Eric Myhre <hash@exultant.us>