Attempt to filter based on ownerReferences.
Unfortunately, I said "attempt".

Which is not to say that the *code* doesn't work -- it does -- but
unfortunately, the clusters in the wild I'm testing against here do not seem
to have many objects correctly following the behavior specified in the docs:
https://kubernetes.io/docs/concepts/workloads/controllers/garbage-collection/

I'm testing against clusters that are version 1.8, so... it's not clear to me
why I'm having a rough time.  Possibly because the clusters contain some
migrated objects?  But I know many of these objects have been created
well after the cluster moved to 1.8, so that's... hrm.

I've only got guesses at this point.  But one way or another, it makes life a
LOT harder.  If we can't tell which objects were auto-created, we're screwed.
Sanity is not possible if you can't enumerate an ownership tree for these.

Signed-off-by: Eric Myhre <hash@exultant.us>